config postgreSQL, typeorm repository and entities, refactoring, version downgrades
