+++ package.json
@@ -23,14 +23,13 @@
     "@nestjs/common": "^9.0.0",
     "@nestjs/core": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
-    "@nestjs/typeorm": "^9.0.1",
+    "@nestjs/typeorm": "8.0.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "pg": "^8.9.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0",
-    "typeorm": "^0.3.12",
-    "uuid": "^9.0.0"
+    "typeorm": "0.2.32"
   },
   "devDependencies": {
     "@nestjs/cli": "^9.0.0",
