+++ src/tasks/dto/get-tasks-filter.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsOptional } from 'class-validator';
-import { TaskStatus } from '../tasks.model';
+import { TaskStatus } from '../task-status.enum';
 
 export class GetTasksFilterDto {
   @IsOptional()
